ref(*): rename services to style guide recs
