+++ src/app/dashboard/dashboard.component.spec.ts
@@ -4,7 +4,7 @@ import { HttpClientTestingModule } from '@angular/common/http/testing';
 
 import { DashboardComponent } from './dashboard.component';
 import { MomentModule } from 'angular2-moment/moment.module';
-import { ProjectService } from '../project.service';
+import { ProjectService } from '../services/project/project.service';
 
 describe('DashboardComponent', () => {
   let component: DashboardComponent;
